pyproject.toml
[project]
name = "adventure-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==5.2",
    "google-genai>=1.59.0",
    "python-dotenv>=1.2.1",
]
